+++ bin/src/extension/string_builder_extension.dart
@@ -1,0 +1,9 @@
+extension StringBuilderExtension on StringBuffer {
+  void writelnWithQuotesOrNull(String key, String? value) {
+    if (value == null) {
+      writeln('        $key: null,');
+    } else {
+      writeln('        $key: \'$value\',');
+    }
+  }
+}