+++ bin/src/model/dto/dependency.dart
@@ -1,15 +1,15 @@
 class Dependency {
   final String name;
-  final String version;
   final String license;
-  final String licenseUrl;
+  final String? version;
+  final String? licenseUrl;
   final String? url;
 
   Dependency({
     required this.name,
-    required this.version,
     required this.license,
-    required this.licenseUrl,
-    required this.url,
+    this.version,
+    this.licenseUrl,
+    this.url,
   });
 }